+++ Gruntfile.js
@@ -13,7 +13,7 @@ module.exports = function (grunt)
                 banner: '/*! <%= "\\r\\n * " + pkg.title %> v<%= pkg.version %> - <%= grunt.template.today("mm/dd/yyyy") + "\\r\\n" %>' +
                     ' * Copyright (c) <%= grunt.template.today("yyyy") %> <%= pkg.author.name %> <%= (pkg.homepage ? "(" + pkg.homepage + ")" : "") + "\\r\\n" %>' +
                     ' * Licensed under <%= pkg.licenses[0].type + " " + pkg.licenses[0].url + "\\r\\n */\\r\\n" %>' + 
-                    ';(function ($, window, undefined)\r\n{\r\n    "use strict";\r\n\r\n',
+                    ';(function ($, window, undefined)\r\n{\r\n    /*jshint validthis: true */\r\n    "use strict";\r\n\r\n',
                 footer: '\r\n})(jQuery, window);',
                 process: function(src, filepath)
                 {